desloppify: split MainViewModel into Groups + Palette partials
MainViewModel.cs sat at 1574 LOC mixing groups/layout, tabs lifecycle,
hydration, palette, status-bar, and wiring concerns. Apply the existing
SidebarViewModel partial pattern: MainViewModel.StatusBar.cs already
covered status-bar; carve out two more concern files so a reader can
predict where each behaviour lives.

- MainViewModel.Groups.cs   — group/layout machinery (Groups,
  FocusedGroup, GroupWidths, GroupWidthsReset, OnGroupsCollectionChanged,
  HookGroupSelection, _pendingSessionToGroup/_pendingFocusedIndex,
  PrepareLayoutFromPersistence, CaptureLayout, CanCloseGroup,
  SplitRight, CloseGroup, FocusGroup/Next/Prev, CycleFocus,
  OpenInNewGroup* commands).
- MainViewModel.Palette.cs  — OpenCommandPaletteAsync + BuildPaletteActions.

Main file shrinks 1574 → 1242 LOC. No public surface changes; the
partial type is the same `public sealed partial class MainViewModel`.

Tests pass; no behavioural change.
